pyproject.toml
[project]
name = "data-monitoring-bot"
version = "0.2.0"
description = "거래소/장비 데이터 수신 감지 + SNMP 수집 + MariaDB 저장 + 휴장일 조회"
requires-python = ">=3.10"
dependencies = [
  "pysnmp>=4.4",
  "mysql-connector-python>=9.0",
  "requests>=2.32",
  "beautifulsoup4>=4.12"
]

[project.scripts]
monitor = "src.main:main"
